refactor: story file moved / components folder created
