+++ my-monorepo/packages/ui/package.json
@@ -4,13 +4,13 @@
   "private": true,
   "exports": {
     ".": "./src/index.ts",
-    "./button/Button": "./src/button/Button.tsx",
-    "./carousel/Carousel": "./src/carousel/Carousel.tsx",
-    "./dropdown/Dropdown": "./src/dropdown/Dropdown.tsx",
-    "./input/Input": "./src/input/Input.tsx",
-    "./modal/Modal": "./src/modal/Modal.tsx",
-    "./sheet/Sheet": "./src/sheet/Sheet.tsx",
-    "./toggle/Toggle": "./src/toggle/Toggle.tsx"
+    "./button/Button": "./src/components/button/Button.tsx",
+    "./carousel/Carousel": "./src/components/carousel/Carousel.tsx",
+    "./dropdown/Dropdown": "./src/components/dropdown/Dropdown.tsx",
+    "./input/Input": "./src/components/input/Input.tsx",
+    "./modal/Modal": "./src/components/modal/Modal.tsx",
+    "./sheet/Sheet": "./src/components/sheet/Sheet.tsx",
+    "./toggle/Toggle": "./src/components/toggle/Toggle.tsx"
   },
   "scripts": {
     "lint": "eslint . --max-warnings 0",
