+++ my-monorepo/packages/ui/src/components/dropdown/Dropdown.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, createContext, useEffect, useRef, useState } from "react";
 import { Button } from "../button/Button";
 import { ButtonProps } from "../button/ButtonType";
-import { useCustomContext } from "../hooks/useCustomContext";
+import { useCustomContext } from "../../hooks/useCustomContext";
 
 const DropdownContext = createContext<{
   open: boolean;
