+++ my-monorepo/packages/ui/src/components/input/input.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Input } from "../Input";
-import { InputProps } from "../Input";
+import { Input } from "./Input";
+import { InputProps } from "./Input";
 import { FieldError } from "react-hook-form";
 import { ChevronDown, LogIn } from "lucide-react";
