+++ my-monorepo/packages/ui/src/components/sheet/Sheet.tsx
@@ -1,11 +1,11 @@
 import { ReactNode, createContext, useRef, useState } from "react";
 import { Button } from "../button/Button";
-import { useOutsideClick } from "../hooks/useOutsideClick";
-import { useDisableScroll } from "../hooks/useDisableScroll";
+import { useOutsideClick } from "../../hooks/useOutsideClick";
+import { useDisableScroll } from "../../hooks/useDisableScroll";
 import { ButtonProps } from "../button/ButtonType";
 import styles from "./Sheet.module.css";
-import { useSheetAnimation } from "../hooks/useSheetAnimation";
-import { useCustomContext } from "../hooks/useCustomContext";
+import { useSheetAnimation } from "../../hooks/useSheetAnimation";
+import { useCustomContext } from "../../hooks/useCustomContext";
 
 export type RootProps = {
   children: ReactNode;
